requirements.txt
Django~=4.0.4
psycopg2-binary==2.9.3
djangorestframework==3.13.1
PyJWT~=2.3.0
numpy~=1.22.4
django-ndarrayfield==1.0.0rc4
Pillow==9.1.1
face-recognition==1.3.0
celery==5.2.6
redis==4.3.1
django-cors-headers==3.12.0